Move HammingToCigarEQX to unused.cpp
